Working version where the app is built out of classes
