+++ lib/Department.js
@@ -7,7 +7,10 @@ class Department {
           name: 'departmentName',
         },
       ]
+      this.viewDepartments = 
+      'SELECT * FROM department;';
     }
+    
   
   }
