+++ lib/Employee.js
@@ -3,6 +3,13 @@ const cTable = require('console.table');    // Import and require console.table
 
 class Employee {
     constructor() {
+      this.viewEmployees = 
+      'SELECT e1.id, e1.first_name, e1.last_name, role.title AS title,'
+      + 'department.name AS department, role.salary AS salary,'
+      + 'CONCAT(e2.first_name," ", e2.last_name) AS manager \nFROM employee AS e1'
+      + '\nJOIN role ON e1.role_id = role.id \nJOIN department ON role.department_id'
+      + '= department.id \nJOIN employee As e2 ON e1.manager_id = e2.id \nORDER BY e1.id;';
+      
       this.addEmployeeQuestions = [
         {
             type: 'input',
@@ -15,13 +22,6 @@ class Employee {
             name: 'employeeLastName',
         },
       ];
-
-      this.viewEmployees = 
-        'SELECT e1.id, e1.first_name, e1.last_name, role.title AS title,'
-        + 'department.name AS department, role.salary AS salary,'
-        + 'CONCAT(e2.first_name," ", e2.last_name) AS manager \nFROM employee AS e1'
-        + '\nJOIN role ON e1.role_id = role.id \nJOIN department ON role.department_id'
-        + '= department.id \nJOIN employee As e2 ON e1.manager_id = e2.id \nORDER BY e1.id;';
     }
   
   }
